Re-configure the stylesheets asset pipeline
This new arrangement imports all base styles into
the two namespaced screen stylesheets, allowing us
to pass global declarations and bootstrap variables
all the way through. It eliminates the pre-compiled
base.css for now, although, it is possible that we will
want to bring that back in the future, if we see a 
need to add vendor styles to both namespaces.

Additionally, this commit tweaks the progress-indicator
styles.