Add the missing template
I just wanted to get us to a point that we were all
green and had a good footing to restart actual development
on this feature. We need to revisit these test again
before moving on. For example, I think the "format review
is incomplete' scope is currently designed to mean
something slightly different than I believe it should 
mean.