Extract this template into a shared partial
Unfortunately, although we need this in both layouts,
it seems that the cocoon gem requires the partial file
to be in the sam directory as the template using it, 
so the only way to use a shared partial is to have the 
local partials point to the shared partial. 
Hopefully the long comment will be enough
to help future developers.